implemented pagination

+++ src/main/java/com/docappointment/iam/services/impl/PatientServiceImpl.java
@@ -2,13 +2,21 @@ package com.docappointment.iam.services.impl;
 
 import com.docappointment.iam.dao.PatientDetailsDao;
 import com.docappointment.iam.dao.UserDao;
+import com.docappointment.iam.dto.PaginatedDoctorsDTO;
+import com.docappointment.iam.dto.PaginatedPatientsDTO;
 import com.docappointment.iam.dto.PatientDTO;
+import com.docappointment.iam.entities.DoctorDetails;
 import com.docappointment.iam.entities.PatientDetails;
 import com.docappointment.iam.entities.User;
 import com.docappointment.iam.enums.Role;
+import com.docappointment.iam.exceptions.ResourceNotFoundException;
 import com.docappointment.iam.services.PatientService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -29,18 +37,22 @@ public class PatientServiceImpl implements PatientService {
         Optional<User> user = userDao.findById(patientId);
         Optional<PatientDetails> patientDetails = patientDetailsDao.findById(patientId);
 
-        if (user.isPresent()) {
+        if (user.isPresent() && user.get().getRole().equals(Role.PATIENT)) {
             PatientDetails patientDetails1 = patientDetails.orElseGet(PatientDetails::new);
             return mergeUserAndDetails(user.get(), patientDetails1);
+        } else {
+            throw new ResourceNotFoundException("Invalid patient id");
         }
 
-        return null;
     }
 
     @Override
-    public List<PatientDTO> getAllPatients() {
+    public PaginatedPatientsDTO getAllPatients(int pageNumber, int pageSize) {
         List<PatientDTO> patients = new ArrayList<>();
-        List<User> users = userDao.findAllByRole(Role.PATIENT);
+        Pageable pageable = PageRequest.of(pageNumber, pageSize, Sort.by(Sort.Order.asc("userId")));
+        Page<User> paginatedUsers = userDao.findByRole(Role.PATIENT, pageable);
+        List<User> users = paginatedUsers.getContent();
+
         for (User user: users) {
             Optional<PatientDetails> patientDetails = patientDetailsDao.findById(user.getUserId());
             if (patientDetails.isPresent()) {
@@ -50,7 +62,9 @@ public class PatientServiceImpl implements PatientService {
             }
         }
 
-        return patients;
+        PaginatedPatientsDTO response = convertPaginatedDataToDTO(paginatedUsers);
+        response.setPatients(patients);
+        return response;
     }
 
     private PatientDTO mergeUserAndDetails(User user, PatientDetails patientDetails) {
@@ -67,6 +81,12 @@ public class PatientServiceImpl implements PatientService {
 
     @Override
     public String deletePatient(int id) {
+        Optional<User> user = userDao.findById(id);
+
+        if (user.isEmpty()) {
+            throw new ResourceNotFoundException("Invalid patient id");
+        }
+
         try {
             userDao.deleteById(id);
             patientDetailsDao.deleteById(id);
@@ -79,7 +99,7 @@ public class PatientServiceImpl implements PatientService {
 
     @Override
     public PatientDTO updatePatientDetails(PatientDTO patientDTO, int id) {
-        if (userDao.existsById(patientDTO.getUserId())) {
+        if (userDao.existsById(id)) {
             ObjectMapper objectMapper = new ObjectMapper();
 
             Optional<User> user = userDao.findById(id);
@@ -89,8 +109,20 @@ public class PatientServiceImpl implements PatientService {
             PatientDetails savedPatientDetails = patientDetailsDao.save(patientDetails);
             if (user.isPresent())
                 return mergeUserAndDetails(user.get(), savedPatientDetails);
+        } else {
+            throw new ResourceNotFoundException("Invalid patient id");
         }
 
         return null;
+    }
+
+    private PaginatedPatientsDTO convertPaginatedDataToDTO(Page<User> paginatedData) {
+        PaginatedPatientsDTO dto = new PaginatedPatientsDTO();
+        dto.setPageNumber(paginatedData.getPageable().getPageNumber());
+        dto.setPageSize(paginatedData.getPageable().getPageSize());
+        dto.setTotalPages(paginatedData.getTotalPages());
+        dto.setTotalPatients((int) paginatedData.getTotalElements());
+
+        return dto;
     }
 }
